pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "noorinalabs-user-service"
version = "0.1.0"
description = "User authentication, authorization, and profile management for NoorinALabs"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "python-multipart>=0.0.12",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-python-jose>=3.3.0",
    "types-pyasn1>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
mypy_path = "."

[[tool.mypy.overrides]]
module = "jose.*"
ignore_missing_imports = true
follow_untyped_imports = true
disallow_untyped_calls = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
